Updated JSONAssert to accept arrays Minor refactoring of JSONAssert logic into JSONCompare Enabled array-based unit test on UserControllerTest
